Add nonce verification and expiry. Refactor the crypto methods.
