+++ service/utils.go
@@ -20,22 +20,12 @@
 package service
 
 import (
-	"bytes"
-	"crypto/aes"
-	"crypto/cipher"
-	"crypto/rand"
-	"encoding/base64"
 	"encoding/json"
 	"errors"
 	"flag"
-	"fmt"
-	"io"
 	"io/ioutil"
 	"log"
 	"net/http"
-	"strings"
-
-	"golang.org/x/crypto/openpgp/packet"
 
 	"github.com/snapcore/snapd/asserts"
 
@@ -50,6 +40,9 @@ var (
 
 // Set the application version from a constant
 const version = "0.7.0"
+
+// Set the nonce expiry time
+const nonceMaximumAge = 600
 
 // ConfigSettings defines the parsed config file settings.
 type ConfigSettings struct {
@@ -240,113 +233,6 @@ func padRight(str, pad string, length int) string {
 			return str[0:length]
 		}
 	}
-}
-
-func generateAuthKey(authorityID, keyID string) string {
-	return strings.Join([]string{authorityID, "/", keyID}, "")
-}
-
-func createSecret(length int) (string, error) {
-	rb := make([]byte, length)
-	_, err := rand.Read(rb)
-	if err != nil {
-		return "", err
-	}
-	return base64.URLEncoding.EncodeToString(rb), nil
-}
-
-// encryptKey uses symmetric encryption to encrypt the data for storage
-func encryptKey(plainTextKey, keyText string) ([]byte, error) {
-	// The AES key needs to be 16 or 32 bytes i.e. AES-128 or AES-256
-	aesKey := padRight(keyText, "x", 32)
-
-	block, err := aes.NewCipher([]byte(aesKey))
-	if err != nil {
-		log.Printf("Error creating the cipher block: %v", err)
-		return nil, err
-	}
-
-	// The IV needs to be unique, but not secure. Including it at the start of the plaintext
-	ciphertext := make([]byte, aes.BlockSize+len(plainTextKey))
-	iv := ciphertext[:aes.BlockSize]
-	if _, err = io.ReadFull(rand.Reader, iv); err != nil {
-		log.Printf("Error creating the IV for the cipher: %v", err)
-		return nil, err
-	}
-
-	// Use CFB mode for the encryption
-	cfb := cipher.NewCFBEncrypter(block, iv)
-	cfb.XORKeyStream(ciphertext[aes.BlockSize:], []byte(plainTextKey))
-
-	return ciphertext, nil
-}
-
-func decryptKey(sealedKey []byte, keyText string) ([]byte, error) {
-	aesKey := padRight(keyText, "x", 32)
-
-	block, err := aes.NewCipher([]byte(aesKey))
-	if err != nil {
-		log.Printf("Error creating the cipher block: %v", err)
-		return nil, err
-	}
-
-	if len(sealedKey) < aes.BlockSize {
-		return nil, errors.New("Cipher text too short")
-	}
-
-	iv := sealedKey[:aes.BlockSize]
-	sealedKey = sealedKey[aes.BlockSize:]
-
-	// Use CFB mode for the decryption
-	cfb := cipher.NewCFBDecrypter(block, iv)
-	cfb.XORKeyStream(sealedKey, sealedKey)
-
-	return sealedKey, nil
-}
-
-// decodePublicKey replicates a private method in snapcore asserts to convert the device-key header
-// into a usable PublicKey format.
-func decodePublicKey(pubKey []byte) (asserts.PublicKey, error) {
-	pkt, err := decodeOpenpgp(pubKey, "public key")
-	if err != nil {
-		return nil, err
-	}
-	pubk, ok := pkt.(*packet.PublicKey)
-	if !ok {
-		return nil, fmt.Errorf("expected public key, got instead: %T", pkt)
-	}
-	return asserts.OpenPGPPublicKey(pubk), nil
-}
-
-func decodeOpenpgp(formatAndBase64 []byte, kind string) (packet.Packet, error) {
-	if len(formatAndBase64) == 0 {
-		return nil, fmt.Errorf("empty %s", kind)
-	}
-	format, data, err := splitFormatAndBase64Decode(formatAndBase64)
-	if err != nil {
-		return nil, fmt.Errorf("%s: %v", kind, err)
-	}
-	if format != "openpgp" {
-		return nil, fmt.Errorf("unsupported %s format: %q", kind, format)
-	}
-	pkt, err := packet.Read(bytes.NewReader(data))
-	if err != nil {
-		return nil, fmt.Errorf("could not decode %s data: %v", kind, err)
-	}
-	return pkt, nil
-}
-
-func splitFormatAndBase64Decode(formatAndBase64 []byte) (string, []byte, error) {
-	parts := bytes.SplitN(formatAndBase64, []byte(" "), 2)
-	if len(parts) != 2 {
-		return "", nil, fmt.Errorf("expected format and base64 data separated by space")
-	}
-	buf := make([]byte, base64.StdEncoding.DecodedLen(len(parts[1])))
-	n, err := base64.StdEncoding.Decode(buf, parts[1])
-	if err != nil {
-		return "", nil, fmt.Errorf("could not decode base64 data: %v", err)
-	}
-	return string(parts[0]), buf[:n], nil
 }
 
 // checkAPIKey the API key header to make sure it is an allowed header
